requirements.txt
torch>=2.0.0
transformers>=4.30.0
biopython>=1.81
numpy>=1.24.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
